Move auto-centering logic into TrajectoryMap; Fix centering only working on first start
